Move remaining code from sct_utils to sct.utils
