+++ spinalcordtoolbox/utils/__init__.py
@@ -1,4 +1,4 @@
 from .fs import *
-from .fs import __version__, __sct_dir__, __data_dir__, __deepseg_dir__
 from .sys import *
+from .sys import __version__, __sct_dir__, __data_dir__, __deepseg_dir__
 from .shell import *
